Require battery techs for ship types that use them
Ship types that use foo_battery_2 should require the tech that enables that battery level, basic_foo_battery, in addition to the hull type; otherwise countries will have ships in the starting OOB with weapons they haven't invented yet, and worse, will be able to create new variants with those weapon types before researching the techs. It appears to be a HoI feature that having a ship variant enables all the equipment variants it uses, ignoring the techs.